added outlet and nested path for product page
