+++ client/src/main.jsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import App from "./App.jsx";
 import ErrorPage from "./errorPage";
-import ProductPage from "./productPage";
+import ProductPage from "./routes/product.jsx";
 import "./index.css";
 
 const router = createBrowserRouter([
