--- client/src/productPage.jsx
@@ -1,5 +1,0 @@
-const ProductPage = () => {
-  return <div>ProductPage</div>;
-};
-
-export default ProductPage;
